refactor(webapp): replace existing Tooltip component
